requirements.txt
requests
beautifulsoup4
python-dotenv
psycopg2-binary
apscheduler
twilio>=8.0.0
slack-sdk>=3.0.0
pyopenssl
cryptography
idna
Flask
Flask-JWT-Extended
Flask-SQLAlchemy
bcrypt
flask-cors
pytz

# Google Chat API dependencies
google-apps-chat>=0.1.9
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
